+++ graphblas_fmm.cpp
@@ -81,7 +81,9 @@ void GraphBLAS_write_FMM(benchmark::State& state) {
     }
     GrB_Matrix_free(&mat);
 
-    std::filesystem::remove(out_path);
+    if (delete_written_files_on_finish) {
+        std::filesystem::remove(out_path);
+    }
     state.SetBytesProcessed((int64_t)num_bytes);
     state.SetLabel("problem_name=" + prob.name);
 }
