+++ pigo_bench.cpp
@@ -71,7 +71,9 @@ static void PIGO_write_binary(benchmark::State& state) {
         benchmark::ClobberMemory();
     }
 
-    std::filesystem::remove(out_path);
+    if (delete_written_files_on_finish) {
+        std::filesystem::remove(out_path);
+    }
 
     state.SetBytesProcessed((int64_t)num_bytes);
     state.SetLabel("problem_name=" + prob.name);
@@ -101,7 +103,9 @@ static void PIGO_write_ascii(benchmark::State& state) {
         benchmark::ClobberMemory();
     }
 
-    std::filesystem::remove(out_path);
+    if (delete_written_files_on_finish) {
+        std::filesystem::remove(out_path);
+    }
 
     state.SetBytesProcessed((int64_t)num_bytes);
     state.SetLabel("problem_name=" + prob.name);
